화면 밖으로 못나감
가장자리에서 화면 밖으로 나갈 수 있는 이동을 선택 할 수 없음

+++ Project/main_state.py
@@ -42,6 +42,8 @@ def handle_events():
             x, y = event.x, 600 - 1 - event.y
             for card in title_state.card_stack:
                 if card.Click(x, y) is True:
+                    if hero.Use_Possible(card.number) is False:
+                        return
                     hero.update_animation(card.number)
                     card.delete()
                     turn += 1
